+++ rust/relay/server/src/main.rs
@@ -747,6 +747,9 @@ where
                 max_elapsed_time,
                 error,
             } => tracing::warn!(?backoff, ?max_elapsed_time, "{error:#}"),
+            Event::RetryAfter { duration, error } => {
+                tracing::warn!(?duration, "Server requested retry-after: {error:#}")
+            }
         }
     }
 }
